reverse the unmasking test
the examples in section 8.8.1 show unmasked to masked, which is what the test
verified. in practice we're going from masked to unmasked, so reverse the test
to reflect that path. in the end it doesn't matter since we're just xoring,
but it feels better using the patterns in the order we'd use them during
decoding.