add broken generic DecodeBCH implementation
this version implements the PGZ algorithm from the BCH wikipedia page, which
should let it work for both format and data decoding. the old version, I now
realize, was usable only for format (the spec trimmed down the math to give
only that which was needed to decode format).

it's broken because
 - it can't handle double bit changes that involve bit zero
 - heaven only knows what it does with 1-, or >2-bit changes
   (the matrix inverter only handles 2x2)